update validation for nsl19m51

+++ STM32/11-modbus-slave-node-1/src/sensors/adc_nsl19m51.h
@@ -7,6 +7,6 @@
 #include <math.h>
 #include "usart2.h"
 
-uint16_t read_NLS19M51_lux(void);
+int32_t read_NLS19M51_lux(void);
 
 #endif /* ADC_NSL19M51_H_ */
